i#1551,i#1569,i#3544 Port sample 'cbr' to AARCHXX and RISCV64.
Modified 1: For RISCV64, conditional branch instruction of  'C' extension may not reach after adding clean call. So like X86, we add support for detect and convert compressed cbr to longer version.
Modified 2: For AARCHXX and RISCV64, a cbr may use the stolen reg and can not be mangled later as it is meta. So we check whether a cbr uses the stolen reg and replace it with a scratch reg.

After thess two modifications, cbr works for AARCHXX and RISCV64.

i#1551,i#1569,i#3544